+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `New-DscResourceWikiPage`
-  - If a class-based resource have a parent class that contains DSC resource
+  - If a class-based resource has a parent class that contains DSC resource
     properties they will now also be returned as part of the DSC resource
     parameters ([issue #62](https://github.com/dsccommunity/DscResource.DocGenerator/issues/62)).